apps/backend/pyproject.toml
[project]
name = "pixels-pygame-palace-backend"
version = "0.1.0"
description = "Flask backend for Pixel's PyGame Palace"
authors = [
    { name = "Jon Bogaty", email = "jon@jonbogaty.com" }
]
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-limiter>=3.13",
    "flask-socketio>=5.5.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "psutil>=7.1.0",
    "pygame-ce>=2.5.5",
    "pygbag>=0.8.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-engineio>=4.12.2",
    "python-socketio>=5.13.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --strict-markers"